+++ src/mplchart/expressions/__init__.py
@@ -4,9 +4,9 @@ from .prelude import wrap_expression, OPEN, HIGH, LOW, CLOSE, VOLUME
 
 from .trend import SMA, EMA, RMA, WMA, HMA, DEMA, TEMA
 
-from .momentum import ROC, MOM, RSI, MACD, STOCH
+from .momentum import ROC, MOM, RSI, PPO, MACD, MACDV, BOP, CMF, MFI, STOCH
 
-from .volatility import TRANGE, ATR, BBANDS, DONCHIAN, KELTNER
+from .volatility import TRANGE, ATR, BBP, BBW, NATR, BBANDS, DONCHIAN, KELTNER, DMI, ADX
 
 from .prices import MIDPRICE, TYPPRICE, WCLPRICE
 
@@ -14,7 +14,7 @@ __all__ = [
     "wrap_expression",
     "OPEN", "HIGH", "LOW", "CLOSE", "VOLUME",
     "SMA", "EMA", "RMA", "WMA", "HMA", "DEMA", "TEMA",
-    "ROC", "MOM", "RSI", "MACD", "STOCH",
-    "TRANGE", "ATR", "BBANDS", "DONCHIAN", "KELTNER",
+    "ROC", "MOM", "RSI", "PPO", "MACD", "MACDV", "BOP", "CMF", "MFI", "STOCH",
+    "TRANGE", "ATR", "BBP", "BBW", "NATR", "BBANDS", "DONCHIAN", "KELTNER", "DMI", "ADX",
     "MIDPRICE", "TYPPRICE", "WCLPRICE",
 ]